backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
PyPDF2
google-generativeai
